requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio] >= 2.0
asyncpg
pydantic
pydantic-settings
elasticsearch[async]==8.13.2
celery >= 5.0
redis
python-dotenv
alembic
psycopg2-binary
requests
faker
uuid
gevent
httpx>=0.27.0